ChatwithPDF/requirements.txt
langchain==0.1.13
PyMuPDF==1.24.0
python-dotenv==1.0.1
streamlit==1.32.2
faiss-cpu==1.8.0
streamlit-extras==0.4.0
openai==1.14.3
tiktoken==0.6.0